backend/requirements.txt
# API Framework
fastapi==0.109.0
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.6.0
pydantic-settings==2.2.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1

# Task Queue
celery==5.3.4
redis==5.0.1

# LLM & AI
google-genai>=1.0.0
tenacity==8.2.3
sentence-transformers==2.2.2

# Document Processing
pdfplumber==0.10.3
python-docx==0.8.11

# Data Processing
pandas==2.1.3
numpy==1.26.2
PyYAML==6.0.1

# Vector Database
qdrant-client>=0.11.0

# PDF Generation
reportlab==4.0.7

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pytz==2023.3

# Logging
python-json-logger==2.0.7

# Multi-Agent & MCP (Option C)
crewai==0.28.8
crewai-tools==0.1.5
